import pre-compiled shaders
the app requires a shader, but producing this shader requires an external
utility called sokol-shdc, which in turn requires 7 git submodules,
and the micro-dependency-centered and cmake-based build tool fips.
apart from that, one of the dependency requires a bleeding edge C++
compiler, targetting a standard much newer than what my GCC 6.5.0
supports.
due to the ridiculous amount of dependencies, the author even provides
a dockerfile to get it built, and the fips description files for this
very repo download a **binary** version of the tool which gets executed
during a build. fortunately i was running the initial test build in
a sandbox, and then used the precompiled tool to generate different
headers with different shader languages.

see https://github.com/floooh/sokol-tools/issues/66